requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
pandas>=2.1.0
numpy>=1.24.0
pyarrow>=13.0.0
regex>=2023.10.3
typing-extensions>=4.7.1

torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.6.0
bitsandbytes>=0.41.0
scikit-learn>=1.3.0
joblib>=1.3.0

faiss-cpu>=1.7.4
rank-bm25>=0.2.2

fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0
openai>=1.14.0

tqdm>=4.66.0
python-dotenv>=1.0.0
huggingface-hub>=0.19.0
cross-encoder>=0.1.6
gradio>=4.0.0
